Avoid bloom index lookup call during record_index bootstrap
Summary: During record index bootstrap the dummy commit creation first creates the record_index and then there will be index lookup during tagging phase. Since, metaclient's table configs are not reloaded it is not finding record_index and it is going through bloom index as a fallback option. This change will reload the table config so that the lookup happens on the record_index. Also, opening up some drogon variables to be overridden by the user in record_index bootstrap drogon app.